update build script
(cherry picked from commit 1254345c5a4ccfa6d062edbe5850b4bb4fae0683)
